test: Extract copy tests to ArithmeticTest file
